+++ cognite/neat/_graph/transformers/_value_type.py
@@ -1,7 +1,15 @@
+import warnings
+from collections.abc import Callable
+from typing import Any, cast
+
+import rdflib
 from rdflib import XSD, Graph, URIRef
 
 from cognite.neat._constants import UNKNOWN_TYPE
 from cognite.neat._graph.queries import Queries
+from cognite.neat._issues.warnings import NeatValueWarning, PropertyDataTypeConversionWarning
+from cognite.neat._utils.auxiliary import string_to_ideal_type
+from cognite.neat._utils.collection_ import iterate_progress_bar
 from cognite.neat._utils.rdf_ import remove_namespace_from_uri
 
 from ._base import BaseTransformer
@@ -64,3 +72,88 @@ class SplitMultiValueProperty(BaseTransformer):
                     graph.remove((s, property_uri, o))
                     new_property = URIRef(f"{property_uri}_{remove_namespace_from_uri(value_type_uri)}")
                     graph.add((s, new_property, o))
+
+
+class ConvertLiteral(BaseTransformer):
+    description: str = "ConvertLiteral is a transformer that improve data typing of a literal value."
+    _use_only_once: bool = False
+    _need_changes = frozenset({})
+
+    _count_by_properties = """SELECT (COUNT(?value) AS ?valueCount)
+    WHERE {{
+      ?instance a <{subject_type}> .
+      ?instance <{subject_predicate}> ?value
+       FILTER(isLiteral(?value))
+    }}"""
+
+    _count_by_properties_uri = """SELECT (COUNT(?value) AS ?valueCount)
+        WHERE {{
+          ?instance a <{subject_type}> .
+          ?instance <{subject_predicate}> ?value
+           FILTER(isIRI(?value))
+        }}"""
+
+    _properties = """SELECT ?instance ?value
+    WHERE {{
+      ?instance a <{subject_type}> .
+      ?instance <{subject_predicate}> ?value
+
+      FILTER(isLiteral(?value))
+
+    }}"""
+
+    def __init__(
+        self,
+        subject_type: URIRef,
+        subject_predicate: URIRef,
+        conversion: Callable[[Any], Any] | None = None,
+    ) -> None:
+        self.subject_type = subject_type
+        self.subject_predicate = subject_predicate
+        self.conversion = conversion or string_to_ideal_type
+        self._type_name = remove_namespace_from_uri(subject_type)
+        self._property_name = remove_namespace_from_uri(subject_predicate)
+
+    def transform(self, graph: Graph) -> None:
+        count_connection_query = self._count_by_properties_uri.format(
+            subject_type=self.subject_type, subject_predicate=self.subject_predicate
+        )
+        connection_count_res = list(graph.query(count_connection_query))
+        connection_count = int(connection_count_res[0][0])  # type: ignore [index, arg-type]
+
+        if connection_count > 0:
+            warnings.warn(
+                NeatValueWarning(
+                    f"Skipping {connection_count} of {self._type_name}.{self._property_name} "
+                    f"as these are connections and not data values."
+                ),
+                stacklevel=2,
+            )
+
+        count_query = self._count_by_properties.format(
+            subject_type=self.subject_type, subject_predicate=self.subject_predicate
+        )
+
+        property_count_res = list(graph.query(count_query))
+        property_count = int(property_count_res[0][0])  # type: ignore [index, arg-type]
+        iterate_query = self._properties.format(
+            subject_type=self.subject_type, subject_predicate=self.subject_predicate
+        )
+
+        for instance, literal in iterate_progress_bar(  # type: ignore[misc]
+            graph.query(iterate_query),
+            total=property_count,
+            description=f"Converting {self._type_name}.{self._property_name}.",
+        ):
+            value = cast(rdflib.Literal, literal).toPython()
+            try:
+                converted_value = self.conversion(value)
+            except Exception as e:
+                warnings.warn(
+                    PropertyDataTypeConversionWarning(str(instance), self._type_name, self._property_name, str(e)),
+                    stacklevel=2,
+                )
+                continue
+
+            graph.add((instance, self.subject_predicate, rdflib.Literal(converted_value)))
+            graph.remove((instance, self.subject_predicate, literal))