+++ docs/CHANGELOG.md
@@ -15,6 +15,10 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Added
+- Support for converting data types with `neat.prepare.instances.convert_data_type(...)`.
+
 ## [0.102.0] - 09-12-**2024**
 ### Fixed
 - Reading data models from CDF ensures unique names for enum collections.